requirements.txt
Django==4.2.8
django-environ==0.9.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
pylint==3.0.3
pylint-django==2.5.5

